+++ application.py
@@ -1,15 +1,15 @@
 from flask import Flask
 
-from api.api import api
-from api.automate_api import automate
-from api.auth_api import auth_api
+from routes.funcx import api
+from routes.automate import automate
+from routes.auth import auth_api
 
 from config import SECRET_KEY
 
 application = Flask(__name__)
 
 # Include the API blueprint
-application.register_blueprint(api, url_prefix="/api/v1")
+application.register_blueprint(api, url_prefix="/routes/v1")
 application.register_blueprint(automate, url_prefix="/automate")
 application.register_blueprint(auth_api)
