+++ routes/funcx.py
@@ -2,13 +2,13 @@ import uuid
 import json
 import time
 
-from .utils import register_endpoint, register_function, get_container, resolve_user, register_container
+from models.utils import register_endpoint, register_function, get_container, resolve_user, register_container
 from authentication.auth import authorize_endpoint, authenticated
 from flask import current_app as app, Blueprint, jsonify, request, abort
 from config import get_redis_client
 
 # Flask
-api = Blueprint("api", __name__)
+api = Blueprint("routes", __name__)
 
 # A cache for authorized endpoint usage by users
 endpoint_cache = {}
